generating unique id

+++ db/db.go
@@ -91,7 +91,7 @@ type Connection struct {
 }
 
 type Record struct {
-	ID        int       `json:"id"`
+	ID        uint64    `json:"id"`
 	Type      EventType `json:"type" gorm:"type:int"`
 	DeviceID  string    `json:"deviceid"`
 	BirthDate int64     `json:"birthdate"`
